+++ cmd/cxchain-cli/main.go
@@ -12,15 +12,13 @@ import (
 var log = logging.MustGetLogger("cxchain-cli")
 
 func main() {
-	initGlobals()
-
 	usageMenu := cxutil.UsageFormat(func(cmd *flag.FlagSet, subcommands []string) {
 		// print: Usage
 		cxutil.PrintCmdUsage(cmd, "Usage", subcommands, []string{"args"})
 
 		// print: ENVs
 		cxutil.CmdPrintf(cmd, "ENVs:")
-		cxutil.CmdPrintf(cmd, "  $%s\n  \t%s", specFileEnv, "chain spec filepath")
+		cxutil.CmdPrintf(cmd, "  $%s\n  \t%s", chainSKEnv, "chain secret key (hex)")
 		cxutil.CmdPrintf(cmd, "  $%s\n  \t%s", genSKEnv, "genesis secret key (hex)")
 
 		// print: Flags
@@ -28,13 +26,16 @@ func main() {
 	})
 
 	root := cxutil.NewCommandMap(flag.CommandLine, 7, usageMenu).
-		AddSubcommand("version", func(_ []string) { cmdVersion() }).
-		AddSubcommand("help", func(_ []string) { flag.CommandLine.Usage() }).
+		AddSubcommand("version", func([]string) { cmdVersion() }).
+		AddSubcommand("help", func([]string) { flag.CommandLine.Usage() }).
 		AddSubcommand("tokenize", cmdTokenize).
 		AddSubcommand("new", cmdNew).
+		AddSubcommand("post", cmdPost).
 		AddSubcommand("run", cmdRun).
 		AddSubcommand("state", cmdState).
-		AddSubcommand("peers", cmdPeers)
+		AddSubcommand("peers", cmdPeers).
+		AddSubcommand("key", cmdKey).
+		AddSubcommand("genesis", cmdGenesis)
 
 	os.Exit(root.ParseAndRun(os.Args[1:]))
 }