+++ CHANGELOG.md
@@ -8,16 +8,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-
 ### Fixed
-
 
 ### Changed
 
-Complete refactor. Preparing for CX Version 1.0 milestone
-
 ### Removed
 
+## [0.76.0] - 2021-03-09
+
+### Added
+
+* Added functionality to locate cx spec via a local file or `cx-tracker` based on a command line flag (https://github.com/skycoin/cx-chains/pull/31/commits/8b3369d981f6b5b6cc7a8007484c1a947c264ba7).
+* Added force client node flag for `cxchain` (https://github.com/skycoin/cx-chains/pull/31/commits/0806031ddc3fd9fb4eebb432ee1cf6bcfd096dfa).
+* Implemented `cxchain-cli post` subcommand (https://github.com/skycoin/cx-chains/pull/31/commits/cd615f99680b6a692eab914d729aa309ddce9810).
+* Implemented `cxchain-cli key` subcommand (https://github.com/skycoin/cx-chains/pull/31/commits/3025f6be7db135a3cf526af51c7901528912eb5f, https://github.com/skycoin/cx-chains/pull/31/commits/ebe2d204516b45ac875c234dff694639e98ebb32)
+* Implemented `cxchain-cli genesis` subcommand (https://github.com/skycoin/cx-chains/pull/31/commits/ebe2d204516b45ac875c234dff694639e98ebb32).
+* Initial work on integration tests (https://github.com/skycoin/cx-chains/pull/31/commits/8b533cb1e6de6bf47047d4f6f2f216cf21fcc9ea, https://github.com/skycoin/cx-chains/pull/31/commits/4dbb0cac096630c9025c0ca1ebc125980624b4ab).
+
+### Fixed
+
+### Changed
+
+* General commandline interface changes for `cxchain` and `cxchain-cli` (https://github.com/skycoin/cx-chains/pull/31/commits/13761d78ea86499d8f3005cc114eca985c68cc1c).
+
+### Removed
 
 ## [0.75.0] - 2020-12-01
 
@@ -27,8 +41,6 @@ Preparing for Version 1.0 milestone. Complete refactor. Massive code changes.
 
 ### Fixed
 
-
 ### Changed
-
 
 ### Removed